Allow IV cache over 251 advances

+++ Core/Gen5/Searchers/IVCacheSearcher.cpp
@@ -130,7 +130,7 @@ void IVCacheSearcher::search(u32 start, u32 end)
         }
 
         RNGList<u8, MT, 32, gen> rngList(seed, initialAdvances);
-        for (u8 i = 0; i <= maxAdvances + 4; i++, rngList.advanceState())
+        for (u32 i = 0; i <= maxAdvances + 4; i++, rngList.advanceState())
         {
             // Entralink
             rngList.advance(22);
